Update restart script to allow rebuilding only at wish

+++ big_tests/restart.sh
@@ -1,3 +1,17 @@
-docker-compose down
-docker-compose build
+print_usage(){
+    echo "Usage:"
+    echo "$0 [rebuild]"
+}
+
+if [[ $# == 1 ]] && [[ "$1" == "rebuild" ]]; then
+    echo "=============== Rebuilding ==============="
+    docker-compose build
+    echo "=========================================="
+elif (( $# >= 1 )); then
+    print_usage
+    exit 1
+else
+    docker-compose down
+    docker-compose rm -f
+fi
 docker-compose up
